feat: remove all functions in backup_common
